+++ src/components/todos/todo.module.css
@@ -1,18 +1,3 @@
-.container {
-  max-width: 1100px;
-  padding: 60px;
-}
-
-.todo-item-container {
-  margin-top: 20px;
-}
-
-.todo-item {
-  display: flex;
-  list-style: none;
-  margin-top: 10px;
-}
-
 .todo-item-title {
   width: 50%;
 }
@@ -31,19 +16,6 @@
   list-style: none;
   gap: 10px;
   align-items: center;
-}
-
-.todo-item-create-container {
-  display: flex;
-  justify-content: flex-end;
-}
-
-.todo-item-create {
-  padding: 0 10px;
-  color: white;
-  border-color: dodgerblue;
-  background-color: dodgerblue;
-  cursor: pointer;
 }
 
 .todo-item-edit {
