+++ src/components/todos/todoList.module.css
@@ -1,0 +1,5 @@
+.todo-item {
+  display: flex;
+  list-style: none;
+  margin-top: 10px;
+}
